+++ src/c/profile/session/read_access.c
@@ -38,7 +38,7 @@ uint16_t mr_write_request_data(mrSession* session, mrStreamId stream_id, mrObjec
     }
 
     // Change this when microcdr supports size_of function.
-    size_t payload_length = 0; //READ_DATA_Payload_size(&payload);
+    uint16_t payload_length = 0; //READ_DATA_Payload_size(&payload);
     payload_length += 4; // (request id + object_id), no padding.
     payload_length += 4; // stream, format, and two optionals.
     payload_length += (control != NULL) ? 8 : 0; // delivery control
